+++ xcube_cci/cciodp.py
@@ -100,18 +100,9 @@ def _convert_time_from_drs_id(time_value: str) -> str:
     return time_value
 
 
-def _run_with_session(async_function, *params):
-    # See https://github.com/aio-libs/aiohttp/blob/master/docs/client_advanced.rst#graceful-shutdown
-    loop = asyncio.get_event_loop()
-    coro = _run_with_session_executor(async_function, *params)
-    result = loop.run_until_complete(coro)
-    # Zero-sleep to allow underlying connections to close
-    loop.run_until_complete(asyncio.sleep(0))
-    return result
-
-
-async def _run_with_session_executor(async_function, *params):
-    async with aiohttp.ClientSession(connector=aiohttp.TCPConnector(limit=50)) as session:
+async def _run_with_session_executor(async_function, *params, headers):
+    async with aiohttp.ClientSession(connector=aiohttp.TCPConnector(limit=50),
+                                     headers=headers) as session:
         return await async_function(session, *params)
 
 
@@ -373,6 +364,7 @@ class CciOdp:
                  num_retries: int = DEFAULT_NUM_RETRIES,
                  retry_backoff_max: int = DEFAULT_RETRY_BACKOFF_MAX,
                  retry_backoff_base: float = DEFAULT_RETRY_BACKOFF_BASE,
+                 user_agent: str = None,
                  only_consider_cube_ready = False
                  ):
         self._opensearch_url = endpoint_url
@@ -381,6 +373,7 @@ class CciOdp:
         self._num_retries = num_retries
         self._retry_backoff_max = retry_backoff_max
         self._retry_backoff_base = retry_backoff_base
+        self._headers={'User_Agent': user_agent} if user_agent else None
         self._drs_ids = None
         self._data_sources = {}
         self._features = {}
@@ -399,9 +392,18 @@ class CciOdp:
     def close(self):
         pass
 
+    def _run_with_session(self, async_function, *params):
+        # See https://github.com/aio-libs/aiohttp/blob/master/docs/client_advanced.rst#graceful-shutdown
+        loop = asyncio.get_event_loop()
+        coro = _run_with_session_executor(async_function, *params, headers=self._headers)
+        result = loop.run_until_complete(coro)
+        # Zero-sleep to allow underlying connections to close
+        loop.run_until_complete(asyncio.sleep(0))
+        return result
+
     @property
     def dataset_names(self) -> List[str]:
-        return _run_with_session(self._fetch_dataset_names)
+        return self._run_with_session(self._fetch_dataset_names)
 
     def get_dataset_info(self, dataset_id: str, dataset_metadata: dict = None) -> dict:
         data_info = {}
@@ -424,7 +426,7 @@ class CciOdp:
 
     def get_datasets_metadata(self, dataset_ids: List[str]) -> List[dict]:
         assert isinstance(dataset_ids, list)
-        _run_with_session(self._ensure_all_info_in_data_sources, dataset_ids)
+        self._run_with_session(self._ensure_all_info_in_data_sources, dataset_ids)
         metadata = []
         for dataset_id in dataset_ids:
             metadata.append(self._data_sources[dataset_id])
@@ -513,7 +515,7 @@ class CciOdp:
         return []
 
     def var_names(self, dataset_name: str) -> List:
-        _run_with_session(self._ensure_all_info_in_data_sources, [dataset_name])
+        self._run_with_session(self._ensure_all_info_in_data_sources, [dataset_name])
         return self._get_data_var_names(self._data_sources[dataset_name])
 
     async def _ensure_all_info_in_data_sources(self, session, dataset_names: List[str]):
@@ -566,7 +568,7 @@ class CciOdp:
         candidate_names = []
         if not self._data_sources and not ecv and not frequency and not processing_level \
                 and not data_type and not product_string and not product_version:
-            _run_with_session(self._read_all_data_sources)
+            self._run_with_session(self._read_all_data_sources)
             candidate_names = self.dataset_names
         else:
             for dataset_name in self.dataset_names:
@@ -596,7 +598,7 @@ class CciOdp:
             converted_start_date = self._get_datetime_from_string(start_date)
         if end_date:
             converted_end_date = self._get_datetime_from_string(end_date)
-        _run_with_session(self._ensure_in_data_sources, candidate_names)
+        self._run_with_session(self._ensure_in_data_sources, candidate_names)
         for candidate_name in candidate_names:
             data_source_info = self._data_sources.get(candidate_name, None)
             if not data_source_info:
@@ -684,7 +686,7 @@ class CciOdp:
                           dimension_names: Dict[str, int],
                           start_time: str = '1900-01-01T00:00:00',
                           end_time: str = '3001-12-31T00:00:00'):
-        dimension_data = _run_with_session(self._get_var_data, dataset_name, dimension_names,
+        dimension_data = self._run_with_session(self._get_var_data, dataset_name, dimension_names,
                                            start_time, end_time)
         return dimension_data
 
@@ -809,7 +811,7 @@ class CciOdp:
 
     def get_time_ranges_from_data(self, dataset_name: str, start_time: str, end_time: str) -> \
             List[Tuple[datetime, datetime]]:
-        return _run_with_session(self._get_time_ranges_from_data,
+        return self._run_with_session(self._get_time_ranges_from_data,
                                  dataset_name,
                                  start_time,
                                  end_time)
@@ -828,7 +830,7 @@ class CciOdp:
         return request_time_ranges
 
     def get_file_list(self, dataset_name):
-        return _run_with_session(self._get_file_list, dataset_name)
+        return self._run_with_session(self._get_file_list, dataset_name)
 
     async def _get_file_list(self, session, dataset_name):
         id = await self._get_dataset_id(session, dataset_name)
@@ -842,7 +844,7 @@ class CciOdp:
         return file_list
 
     def get_dataset_id(self, dataset_name: str) -> str:
-        return _run_with_session(self._get_dataset_id, dataset_name)
+        return self._run_with_session(self._get_dataset_id, dataset_name)
 
     async def _get_dataset_id(self, session, dataset_name: str) -> str:
         await self._ensure_in_data_sources(session, [dataset_name])
@@ -857,7 +859,7 @@ class CciOdp:
         return feature_list[0][2]
 
     def get_data_chunk(self, request: Dict, dim_indexes: Tuple) -> Optional[bytes]:
-        data_chunk = _run_with_session(self._get_data_chunk, request, dim_indexes)
+        data_chunk = self._run_with_session(self._get_data_chunk, request, dim_indexes)
         return data_chunk
 
     async def _get_data_chunk(self, session, request: Dict, dim_indexes: Tuple) -> Optional[bytes]:
@@ -1115,7 +1117,7 @@ class CciOdp:
         return variable_infos, dataset.attributes
 
     def get_opendap_dataset(self, url: str):
-        return _run_with_session(self._get_opendap_dataset, url)
+        return self._run_with_session(self._get_opendap_dataset, url)
 
     async def _get_opendap_dataset(self, session, url: str):
         if url in self._datasets: